Split class I affinity predictor helpers
