+++ mhcflurry/affinity/__init__.py
@@ -10,4 +10,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-__version__ = "2.3.0rc11"
+
+"""Helper modules for class I affinity prediction."""
